+++ BrainPortal/cbrain_plugins/cbrain-plugins-base/cbrain_task/parallelizer/bourreau/parallelizer.rb
@@ -61,8 +61,8 @@ class CbrainTask::Parallelizer < ClusterTask #:nodoc:
     subtasks.each do |otask|
       odir    = otask.full_cluster_workdir
       oscript = otask.qsub_script_basename
-      oout    = otask.stdout_cluster_filename
-      oerr    = otask.stderr_cluster_filename
+      oout    = otask.qsub_stdout_basename
+      oerr    = otask.qsub_stderr_basename
       commands += [
         "",
         "# Run task #{otask.fullname}",