+++ Release-Notes.md
@@ -1,5 +1,22 @@
 ## CBRAIN Release Notes
 
+#### Version 4.6.1 Released 2016-12-19
+
+- A much more complete version of the Swagger API spec (but not final).
+- Support for version 0.4 of the Boutiques descriptors.
+- Misc bug fixes.
+
+#### Version 4.6.0 Released 2016-11-21
+
+- Build-in support for Travis Continuous Integration.
+- User registration form.
+- Preliminary Swagger API support (incomplete!).
+- New credits page.
+- Browsing files within a collection now allows downloading them.
+- Boutique support for containers.
+- Many UI improvements.
+- Many internal bug fixes and performance enhancements.
+
 #### Version 4.5.0 Released 2016-08-18
 
 - Preliminary (alpha) support for Amazon clusters.